Remove AnonymousIdentifiable protocol
